pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tiktok-ads-mcp"
version = "0.1.3"
authors = [
  { name = "Yu Shengnan", email = "yushengnan0525@gmail.com" },
]
description = "Model Context Protocol (MCP) plugin for interacting with TikTok Ads API"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["tiktok", "ads", "mcp", "api", "marketing", "advertising", "model-context-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "requests>=2.32.3",
    "python-dotenv>=1.0.0",
    "mcp>=1.9.0",
    "pandas>=2.2.0",
    "httpx>=0.27.0",
    "tenacity>=8.2.0",
    "google-cloud-firestore>=2.16.0",
    "uvicorn>=0.30.0",
]

[project.scripts]
tiktok-ads-mcp = "tiktok_ads_mcp.main:cli"

[project.urls]
Homepage = "https://github.com/ysntony/tiktok-ads-mcp"
Repository = "https://github.com/ysntony/tiktok-ads-mcp"
Documentation = "https://github.com/ysntony/tiktok-ads-mcp#readme"
"Bug Tracker" = "https://github.com/ysntony/tiktok-ads-mcp/issues"
"Changelog" = "https://github.com/ysntony/tiktok-ads-mcp/blob/main/README.md#changelog"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "twine>=4.0.0",
    "build>=1.0.0",
]

http = [
    "uvicorn>=0.30.0",
    "starlette>=0.40.0",
]

[tool.hatch.build.targets.wheel]
packages = ["tiktok_ads_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "/tiktok_ads_mcp",
    "/README.md",
    "/LICENSE",
    "/MCP_USAGE.md",
    "/pyproject.toml",
]